fix: 로딩 중 텍스트 추가 및 useSSE 복원
