+++ src/Domain/User/components/MyTable.jsx
@@ -1,83 +1,4 @@
-import api from "../../../utils/api";
-import useSSE from "../../../hooks/useSSE";
-import { useEffect, useState } from "react";
-
-const MyTable = () => {
-  const [stocks, setStocks] = useState([]);
-  const sseData = useSSE("/subscribe/portfolio-price");
-
-  useEffect(() => {
-    const fetchStockData = async () => {
-      try {
-        const response = await api.get("/users/stocks");
-
-        const closingPriceResponse = await api.get(
-          "/users/stocks/closing-price"
-        );
-
-        const updatedStocks = response.data.map((stock) => {
-          const closingStock = closingPriceResponse.data.find(
-            (s) => s.stockCode === stock.stockCode
-          );
-
-          if (closingStock) {
-            const totalValue = stock.amount * closingStock.closingPrice;
-            const totalReturn =
-              ((closingStock.closingPrice - stock.averagePrice) /
-                stock.averagePrice) *
-              100;
-
-            return {
-              ...stock,
-              closingPrice: closingStock.closingPrice,
-              totalValue,
-              totalReturn: totalReturn.toFixed(2),
-            };
-          }
-
-          return {
-            ...stock,
-            closingPrice: null,
-            totalValue: stock.amount * stock.averagePrice,
-            totalReturn: null,
-          };
-        });
-
-        setStocks(updatedStocks);
-      } catch (error) {
-        console.error(
-          "데이터 불러오기 실패:",
-          error.response?.data || error.message
-        );
-      }
-    };
-
-    fetchStockData();
-  }, []);
-
-  useEffect(() => {
-    if (!sseData) return;
-
-    setStocks((prevStocks) =>
-      prevStocks.map((stock) => {
-        if (stock.stockCode === sseData.stockCode) {
-          const currentPrice = Number(sseData.currentPrice);
-          const totalValue = stock.amount * currentPrice;
-          const purchasePrice = stock.amount * stock.averagePrice;
-          const totalReturn =
-            ((totalValue - purchasePrice) / purchasePrice) * 100;
-
-          return {
-            ...stock,
-            totalValue,
-            totalReturn: totalReturn.toFixed(2),
-          };
-        }
-        return stock;
-      })
-    );
-  }, [sseData]);
-
+const MyTable = ({ stocks }) => {
   return (
     <div className="space-y-3 py-5 bg-white p-3 w-full max-w-xs">
       {stocks.length > 0 ? (
