+++ src/hooks/useSSE.js
@@ -7,12 +7,9 @@ const BASE_SSE_URL = `http://${import.meta.env.VITE_API_BASE_URL}:${
   import.meta.env.VITE_API_BASE_PORT
 }/sse`;
 
-const RETRY_DELAYS = [1000, 3000, 5000, 10000];
-
 const useSSE = (endpoint) => {
   const [data, setData] = useState(null);
   const { token, setToken } = useAuth();
-  const [retryCount, setRetryCount] = useState(0);
   const controller = new AbortController();
 
   useEffect(() => {
@@ -22,8 +19,7 @@ const useSSE = (endpoint) => {
     const sseUrl = `${BASE_SSE_URL}${endpoint}`;
 
     const connect = () => {
-      if (isCancelled || retryCount >= RETRY_DELAYS.length) return;
-
+      if (isCancelled) return;
       setToken(localStorage.getItem("Authorization"));
 
       fetchEventSource(sseUrl, {
@@ -35,52 +31,32 @@ const useSSE = (endpoint) => {
         signal: controller.signal,
         onopen(response) {
           if (response.status === HttpStatusCode.Forbidden) {
-            setToken(null);
-            console.error("로그인 실패");
+            console.log("로그인이 필요합니다.");
           }
+          console.log(`SSE 연결 성공: ${sseUrl}`, response);
         },
         onmessage(event) {
+          console.log(`SSE 데이터 수신 (${sseUrl}):`, event.data);
           let parsedData;
           try {
             parsedData = JSON.parse(event.data);
-            setRetryCount(0);
           } catch (error) {
             parsedData = event.data;
             console.error("SSE 데이터 파싱 실패:", error);
-
-            if (retryCount < RETRY_DELAYS.length) {
-              const delay = RETRY_DELAYS[retryCount];
-              console.warn(`⏳ ${delay / 1000}초 후 재시도...`);
-              setTimeout(connect, delay);
-              setRetryCount(retryCount + 1);
-            } else {
-              console.error("SSE 데이터 파싱 실패 - 최대 재시도 횟수 도달");
-              controller.abort();
-            }
           }
           setData(parsedData);
         },
         onerror(error) {
           console.error(`SSE 오류 발생 (${sseUrl}):`, error);
           controller.abort();
-
-          if (!isCancelled && retryCount < RETRY_DELAYS.length) {
-            const delay = RETRY_DELAYS[retryCount];
-            console.warn(`⏳ ${delay / 1000}초 후 SSE 재연결...`);
-            setTimeout(connect, delay);
-            setRetryCount(retryCount + 1);
-          } else {
-            console.error("SSE 연결 실패 - 최대 재시도 횟수 도달");
+          if (!isCancelled) {
+            setTimeout(connect, 3000);
           }
         },
         onclose() {
-          if (!isCancelled && retryCount < RETRY_DELAYS.length) {
-            const delay = RETRY_DELAYS[retryCount];
-            console.warn(`⏳ ${delay / 1000}초 후 SSE 재연결...`);
-            setTimeout(connect, delay);
-            setRetryCount(retryCount + 1);
-          } else {
-            console.error("SSE 연결 종료 - 최대 재시도 횟수 도달");
+          console.log(`SSE 연결 종료 (${sseUrl})`);
+          if (!isCancelled) {
+            setTimeout(connect, 3000);
           }
         },
       });
@@ -89,11 +65,11 @@ const useSSE = (endpoint) => {
     connect();
 
     return () => {
-      console.log(`🔌 SSE 연결 cleanup (${sseUrl})`);
+      console.log(`SSE 연결 cleanup (${sseUrl})`);
       isCancelled = true;
       controller.abort();
     };
-  }, [endpoint, token, retryCount]);
+  }, [endpoint, token]);
 
   return data;
 };
